pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
